Custom table navigation extracted to the separate plugin. Navigation fixes.
