+++ src/table.js
@@ -13,6 +13,7 @@ import TableEditing from './tableediting';
 import TableUI from './tableui';
 import TableSelection from './tableselection';
 import TableClipboard from './tableclipboard';
+import TableNavigation from './tablenavigation';
 import Widget from '@ckeditor/ckeditor5-widget/src/widget';
 
 import '../theme/table.css';
@@ -36,7 +37,7 @@ export default class Table extends Plugin {
 	 * @inheritDoc
 	 */
 	static get requires() {
-		return [ TableEditing, TableUI, TableSelection, TableClipboard, Widget ];
+		return [ TableEditing, TableUI, TableSelection, TableClipboard, TableNavigation, Widget ];
 	}
 
 	/**
